function that prints 10 times the numbers, from 0 to 14, followed by a new line.

+++ 0x04-more_functions_nested_loops/5-more_numbers.c
@@ -8,7 +8,7 @@
 
 void more_numbers(void)
 {
-long ch;
+int ch;
 int i = 0;
 while (i < 10)
 {
